Sanitize quotes for wpa_start
Arguments to wpa_start shouldn't be expanded until wpa_start receives
them: so should be quoted by all calling functions. I added quotes
throughout. In general, it seems wise to quote except when it's
absolutely certain not to be needed---or you specifically _do_ want
to expand arguments. In this case, it's extra important.

Because this might be missed in later code, I also wrangled wpa_start
so that it doesn't rely on WPA_OPTS being all in $3.

One of these fixes would have sufficed, but what the hell I did both.

Signed-off-by: Jim Pryor <profjim@jimpryor.net>