network: is_interface ignoring $1
Was the intent for is_interface to be passed the INTERFACE as a
parameter (that's how it's called in most or all places)? Or to read
INTERFACE from the current environment? That's how it's implemented.

Also made INTERFACE in bring_interface a local variable. I don't think
this should break anything, but I'm only 95% confident.