systemd: wait for network to come up
We need to make sure no the systemd services are only considered
started when the network is actually up. Otherwise, things relying
on the network don't know when they can start.

This means that we need to use Type=forking or Type=oneshot, and in
the case of ifplugd we must daemonize and wait for the daemon to be
ready before we continue.

Also do some minor cleanup:
 - don't include env files that are not used
 - don't ignore return values

Signed-off-by: Tom Gundersen <teg@jklm.no>