Fix bond connections (FS#31770)
This commit removes an undocumented feature where IFACE_{UP,DOWN} would
be evaluated when the interface is brought up or down.
This functionality did not work as expected since the interface could
potentially be brought up/down more than once in one netcfg invocation.
This is apparent in the changes to the bonding code.