only grab the profile name for zsh completion
Instead of grabbing the whole unit file, it should only grab the profile
name. Also, make sure it exits after the number of possible commands is
full.

Since there can only be one argument after one of the commands, make
that into a rule.

Include logic so that incorrect values cannot be based based on if the
profile is already started/stopped, or enabled/disabled.