[wifi-menu] Autodetect network interfaces
After the systemd 197 device name changes, wlan0 no longer exists (it
does for people who mask the rule, but new users will have issues with
this). This just searches the /sys/class/net directory for wireless
names and echoes them. If there is more than one, then report an
"Invalid interface specification", not a "Missing" interface (as there
could be multiple).

A quick thanks to Dave Reisner for his help in the channel.

Signed-off-by: William Giokas <1007380@gmail.com>