[wifi-menu] Do not fail if already connected
If netctl is already connected to a network and wifi-menu is run, then
it will fail to connect. Having wifi-menu call `netctl switch-to`
instead of `netctl restart` causes wifi-menu to behave like netctl
switch-to: It will connect if not connected, and if already connected on
that interface it will put the connected profile down the put up the
selected profile.

This fixes github issue #6.

Signed-off-by: William Giokas <1007380@gmail.com>