Altered PKGBUILD to depend on dhcpcd, wpa_actiond to ensure that if installed explicitly it has all necessary components for the user to properly configure and test his connection.
Added netctl.install to notify the customer to review netctl.profile(5),
netctl(1), and netctl.special(7) for details on how to configure their
network profiles.

I feel like echo'ing this should take care of referencing man pages in
any of the examples. Users should view output of makepkg and refference
the manual pages which will tell them where the profiles are stored as
well as how to use options available to them.