Uniform translation of profiles to unit names
The unit corresponding to a profile named $p is netctl@$p.service.
We should not drop the suffix, since $p could contain a valid unit
suffix (although .service is not allowed).
We take care of this uniformly by wrapping around systemctl.
Additionally, `systemctl list-units` does not accept an '--active'
parameter, so we fix it to be consistent with `netctl list`.