Allow group-accessible tuntap devices
The tuntap connection type requires both a User and a Group to be
defined. However, when configuring a tuntap device in the kernel, the
device's uid and gid must both match for someone to be able to use the
device. The only way to create a tuntap device accessible to a group of
users is therefore to avoid specifying the user argument to
"ip tuntap add".

Reported by: A Web