Properly check existence of a connection type.
Substituting "/dev/null/nonexistent" for an empty $Connection,
of course, does generally work, but is somewhat ugly (esp. in that
the string is appended to $CONN_DIR, so the resulting path may
theoretically exist).

Instead, substitute an empty string for the complete path if $Connection
is empty: an empty string can never be readable.

Signed-off-by: Ivan Shapovalov <intelfx100@gmail.com>