Make netctl-auto more hook friendly
The hooks (including any interface hook)  are loaded each time a wpa
event occurs. This means the auto.action script can be extended in the
hooks. For this purpose, several exported variables (mainly $ACTION)
can be checked for their value. The name of the active profile is not
exported, but still available (just as in ordinary netctl usage) through
the $Profile variable.

Profiles ending in .action are disallowed (reserved for possible future
use). Both ifplugd and wpa_actiond have a notion of 'action script'.