Prohibit user invocation of netctl-auto [start|stop]
Instead of using systemctl to manage the netctl-auto service, some users try to
invoke the start/stop commands of netctl-auto directly. Although this behaviour
is explicitly discouraged in the man page, this seems to be a source of
confusion.

The changes introduced in 537aae will cause other netctl-auto commands to fail
if the netctl-auto service was not spawned by systemd. Therefore, this change
emits an error message if netctl-auto [start|stop] is invoked from an
interactive terminal.