Introduce a new PPP connection type
This should at some point take care of most PPP use cases, but
currently is just a frontend to pppoe. The main difference is that the
new connection type keeps track of profiles by their 'dynamic' interface
name. Hence it is not wise to use both pppoe and ppp profiles.