Distinguish between network{,-online}.target (FS#50476)
The netctl@ service now yields before obtaining an IP address. This
prevents holding up network.target unnecessarily. Waiting for profiles
to obtain an IP address is possible through

1) The netctl-wait-online service
This service waits until all enabled profiles have obtained an address.
It is ordered before network-online.target so that this target is now
correctly implemented in netctl.

2) The wait-online <PROFILE> subcommand to netctl
This command waits for a started profile to obtain an address.