Add descriptive solution to netctl-auto error
The "$INTERFACE is controlled by netctl-auto" error results after running wifi-menu, creating a new profile, and failing to alert netctl of the new profile.

This approach works every time for my system, to the point that you may even want to do it automatically, and print that the script is trying to reconnect via netctl-auto. 

In lieu of any automatic solutions to this minor issue, it would be better to have a more descriptive error.