dhcp/dhcpcd: DhcpcdOptions: Don't forcefully append `-L` (`--noipv4ll`).
While it might be desirable to have it disable IPv4LL by default, the
current implementation unconditionally forcefully disabled it, with no way
to enable it.

At the end of the day, the tool was trying to be smarter than the user.
The antithesis of KISS.

Now, one might balk that this changed enables IPv4LL by default.  Except it
still doesn't.  IPv4LL is disabled in the default dhcpcd.conf (at least as
shipped by Arch Linux).  If this change makes IPv4LL be on by default for
someone, it's because that person already took manual steps to enable it.