Do not forcibly disable IPv4LL with dhcpcd
Turn "-L" into a default value of DhcpcdOption=, making it possible to forego
the parameter by setting the variable to an empty string.

DhcpcdOptions= is currently not documented. The last documentation read:
"Do not use this unless you know what you are doing."
When using this variable, "-L" now has to be added if its functionality is
desired.