Makefile: specify the correct shell to use
The Makefile commands in use include bash-specific features like brace
expansion, and thus cannot run under the default /bin/sh shell as that
may or may not be GNU bash (bash supports brace expansion even in posix
mode).

Consequently, the Makefile failed to execute correctly when the system
/bin/sh was a different sh implementation, such as dash or busybox ash.

Fix this by declaring the SHELL which will be used when running all
commands.