Speed up 'netctl list'
When you call `netctl list`, it calls `systemctl status` for each of the profiles. Then it `sed`s out just a single `Status` field.
But when you call `systemctl status`, it fetches last log entries from `journald` which is often time-consuming operation.

So we should pass `--lines=0` which effectively tells `systemctl` to not ask `journalctl`.