Don't delay deconfiguring in netctl-ifplugd (FS#67382)
When netctl-ifplugd starts a profile that uses dhcpcd and the interface
disappears (for example because of a re-enumeration after a resume from
suspend), dhcpcd immediately exits, but ifplugd notices that the interface
reappears within a timeout and does not bother reconfiguring it.

This commit disables the ifplugd timeout. If it turns out to be needed for
buggy interfaces, we should implement the desired functionality in the
ifplugd.action script.