Add verify subcommand to netctl
The name of the subcommand mirrors the same subcommand of systemd-analyze.

Verification before restarting would be useful. However, verification
requires the profile file to be readable by the current user. The current
user may be a regular user with permission to control systemd, so
restarting does not require the profile file to be readable. Therefore, we
do not attempt verification before restarting.